emacs: allow generating passwords in place
Previously, generating a password would `--force` over the entry and
replace the whole file with a new password.

Amend this behaviour by detecting whether the entry exists and
leveraging pass's `--in-place` flag to only replace the password (i.e
the top line). This requires a bit of logic because `--force` and
`--in-place` are mutally exclusive.

Also remove the optional arguments to `password-store--run-generate`
as they were only used, arbitrarily, from `password-store-generate`.